Add environment Add main file Finish simulation process

+++ src/random_variables_simulation.py
@@ -4,9 +4,9 @@ import math
 def uniform(a, b):
     U = rnd.random()
 
-    return a + ((b - a)*U)
+    return int(a + ((b - a)*U))
 
 def exponential(h):
     U = rnd.random()
     
-    return -1 * (math.log(U) * h)
+    return int(-1 * (math.log(U) * h))
